pyproject.toml
[project]
name = "motrpac-backend-utils"
version = "0.9.4"
description = "Common utilities for MoTrPAC services"
authors = [{ name = "Mihir Samdarshi", email = "msamdars@stanford.edu" }]
license = "Apache-2.0"
requires-python = ">=3.11,<3.14"

dependencies = [
    "google-auth",
    "google-cloud-logging",
    "opentelemetry-api",
    "opentelemetry-exporter-gcp-trace>=1.7.0",
    "opentelemetry-instrumentation-logging",
    "opentelemetry-instrumentation-threading",
    "opentelemetry-instrumentation-urllib3",
    "opentelemetry-propagator-gcp",
    "opentelemetry-sdk",
    "typing-extensions",
]

[project.optional-dependencies]
zipper = [
    "google-cloud-storage",
    "google-cloud-pubsub",
    "psutil",
    "smart-open",
    "protobuf~=6.33",
    "pydantic>=2.0",
]
messaging = ["google-cloud-pubsub", "protobuf~=6.33", "pydantic>=2.0"]
flask = ["flask"]

[dependency-groups]
dev = ["ruff"]
test = ["pytest~=8.4.2", "pytest-mock~=3.14",  "pytest-cov~=7.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

# Which packages we want to build/distribute.
# There are also ways of including additional files, such as data files.
# See the Hatch documentation for more information.
[tool.hatch.build.targets.wheel]
packages = ["src/motrpac_backend_utils"]

# Pytest is a testing framework for Python.
[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    "-ra", "--full-trace",
    "--import-mode=importlib",
]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 96
target-version = "py311"
extend-exclude = ["*_pb2.py", "*_pb2.pyi"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "ANN401",
    "D203",
    "D212",
    "D401"
]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["D100", "D101", "D102", "D103", "INP001", "S101", "S108", "SLF001"]
